+++ eRecruiter.Api.Client/Requests/ApplicantCompass/InterestProfileRequest.cs
@@ -1,0 +1,16 @@
+﻿using eRecruiter.Api.Responses;
+using System.Net.Http;
+
+namespace eRecruiter.Api.Client.Requests
+{
+    public class InterestProfileRequest : HttpRequestMessage<InterestProfileResponse>
+    {
+        /// <summary>
+        /// Fetch an interest profile by its ID
+        /// </summary>
+        /// <param name="interestProfileId">Interest Profile ID</param>
+        public InterestProfileRequest(int interestProfileId) : base(HttpMethod.Get, $"/ApplicantCompass/InterestProfile/{interestProfileId}")
+        {
+        }
+    }
+}