+++ eRecruiter.Api/Responses/ApplicantCompass/JobInsightTask.cs
@@ -1,0 +1,12 @@
+﻿namespace eRecruiter.Api.Responses
+{
+    /// <summary>
+    /// Job Insight Task
+    /// </summary>
+    public class JobInsightTask
+    {
+        public string Title { get; set; }
+        public string Text { get; set; }
+        public byte[] Image { get; set; }
+    }
+}